alien base class, biological base class, and mechanical base classes completed. Smaller entity refactor to new folder. Slight imporvment to how shooting works
