+++ src/main/java/TK/Entities/AlienStuff/Aliens/BioMechanical.java
@@ -1,0 +1,5 @@
+package TK.Entities.AlienStuff.Aliens;
+
+public enum BioMechanical {
+    BIOLOGICAL, MECHANICAL;
+}
