+++ src/main/java/TK/Entities/Weapons/Range.java
@@ -1,0 +1,7 @@
+package TK.Entities.Weapons;
+
+public enum Range {
+    // not implemented yet but this will dictate the effective
+    // firing range in tiles once the UI is not implemented
+    MELEE, LOW, MEDIUM, HIGH
+}
