--- src/main/java/TK/game/EntityStuff/AlienStuff/AlienType.java
@@ -1,6 +1,0 @@
-package TK.game.EntityStuff.AlienStuff;
-
-public enum AlienType {
-    SECTOID, DRONE, THINMAN, FLOATER, OUTSIDER, CHRYSSALID, MUTON, HEAVYFLOATER, CYBERDISC, MUTONELITE, BERSERKER,
-    SECTOIDCOMMANDER, ETHEREAL, SECTOPOD, UBERETHEREAL;
-}
