--- src/main/java/TK/game/Weapons/Range.java
@@ -1,7 +1,0 @@
-package TK.game.Weapons;
-
-public enum Range {
-    // not implemented yet but this will dictate the effective
-    // firing range in tiles once the UI is not implemented
-    MELEE, LOW, MEDIUM, HIGH
-}
