--- src/main/java/TK/game/Weapons/WeaponMods/LaserSight.java
@@ -1,22 +1,0 @@
-package TK.game.Weapons.WeaponMods;
-
-public class LaserSight extends WeaponMod {
-    public LaserSight(ModTier tier) {
-        super(tier);
-    }
-
-    @Override
-    public Integer applyEffect() {
-        return (this.tier.ordinal() + 1) * 5; // 5-10-15
-    }
-
-    @Override
-    public String getName() {
-        return "Laser Sight";
-    }
-
-    @Override
-    public String getDescription() {
-        return "Increases critical chance";
-    }
-}
