--- src/main/java/TK/game/Weapons/WeaponMods/ModTier.java
@@ -1,15 +1,0 @@
-package TK.game.Weapons.WeaponMods;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public enum ModTier {
-    BASIC, ADVANCED, SUPERIOR;
-
-    public static final Map<ModTier, String> modTierNameMap = new HashMap<>();
-    static {
-        modTierNameMap.put(BASIC, "Basic");
-        modTierNameMap.put(ADVANCED, "Advanced");
-        modTierNameMap.put(SUPERIOR, "Superior");
-    }
-}
