--- src/main/java/TK/game/Weapons/WeaponMods/Stock.java
@@ -1,22 +1,0 @@
-package TK.game.Weapons.WeaponMods;
-
-public class Stock extends WeaponMod{
-    public Stock(ModTier tier) {
-        super(tier);
-    }
-
-    @Override
-    public Integer applyEffect() {
-        return this.tier.ordinal() + 1; // 1-2-3
-    }
-
-    @Override
-    public String getName() {
-        return "Stock";
-    }
-
-    @Override
-    public String getDescription() {
-        return "Missed shots deal damage";
-    }
-}
