--- src/main/java/TK/game/Weapons/WeaponTier.java
@@ -1,13 +1,0 @@
-package TK.game.Weapons;
-
-import java.util.HashMap;
-import java.util.Map;
-
-public enum WeaponTier {
-    CONVENTIONAL, MAGNETIC, PLASMA;
-    public static final Map<WeaponTier, WeaponTier> tierMap = new HashMap<>();
-    static {
-        tierMap.put(CONVENTIONAL, MAGNETIC);
-        tierMap.put(MAGNETIC, PLASMA);
-    }
-}
