+++ src/main/java/TK/game/game.java
@@ -7,19 +7,7 @@ import java.util.Random;
 
 public class game {
     static Random rd = new Random();
-    public static int getRandomIntInRange(int min, int max) {
-        if (min > max) {
-            throw new IllegalArgumentException("Minimum cannot be higher than maximum");
-        }
-        return rd.nextInt((max - min) + 1) + min;
-    }
 
-    public static String getRandomListItem(List<String> list) {
-        if (!(list instanceof List<String>)) {
-            throw new IllegalArgumentException("Wrong list type");
-        }
-        return list.get(rd.nextInt(list.size()));
-    }
 
     public static void createMap(){
         int totalAmountOfRooms = getRandomIntInRange(50, 100);
