+++ src/main/java/TK/ui/createWindow.java
@@ -1,4 +1,11 @@
 package TK.ui;
 
-public class createWindow {
+import javafx.application.Application;
+import javafx.stage.Stage;
+
+public class createWindow extends Application {
+    @Override
+    public void start(Stage primaryStage) throws Exception {
+
+    }
 }
